Створив окремий файл custom_inline_keyboard.py для користувацьких клавіатур та змінив зв'язки імпорту
